refactor(kernelforge): remove four superseded subsystems (-3.4k production lines)
Each of these had been replaced in place rather than deleted, so the old
generation kept shipping with no reader:

- `gemm_tune/tier3/` (10 files). The gate `warrants_generated_tuner` fires
  only for tables the dispatcher has no entry for, and the dispatcher admits
  exactly one table -- the two predicates accept disjoint sets. On the path
  where the gate does fire, the runner discards the model-authored tuner
  unconditionally at the referee stage. The live coverage reporting moved to
  `gemm_tune/coverage.py` with the dead property dropped; its tests moved to
  `test_coverage_gaps.py`.

- `learning/` (4 files). It wrote through `tuning_db.py`, whose
  `_TUNING_DB_WRITE_ENABLED` has been `False` since
  `knowledge/experience_sink.py` took over, and the output files had no
  reader. `IterationLoop`'s `evolver` parameter and its two call sites go
  with it.

- The tracker's scoring layer (`best_iteration`, `summary_table`,
  `KernelScoringView` and friends in `tracker/schema.py`, `get_best` /
  `summary` in `tracker/experiment.py`). This is the pre-`loop/scoring.py`
  scorer; production reads only `.iterations` and `.checkpoint` off the
  tracker, and `loop/runner.py` carries its own `_is_gate_met` and
  `best_mean_case_speedup`. `list_experiments` is kept -- it is an
  independent state-inspection helper.

- The seven-function reachability island in `fusion/validate.py`
  (`unreached_fusion_symbols` and its private helpers), whose only references
  were each other's definitions. `fused_symbol_invocation_evidence`, which
  `fusion/command.py` does call, is untouched.

Also in this change:

- Drop the deprecated `kernel-agents` console script. The rename shipped in
  v1.0.0b2 and the alias was kept for one release; nothing in the repository,
  the docs, or the examples invoked it. The `kernel_agents.agent_providers`
  entry-point group stays -- it is how pre-rename third-party provider plugins
  are discovered, and it is not a CLI surface.

- Factor the newline-delimited JSON-RPC envelope shared by the two MCP stdio
  servers into `mcp_server/stdio_transport.py`. The two copies were
  byte-identical, only one was tested, and `InvalidParamsError` existed twice
  so an `isinstance` check could not span them.

- De-duplicate test helpers that had drifted into three or five copies:
  `_git` / `_GIT_IDENTITY` across the controller tests, the KB Store run
  config across three modules, and the agent-provider registry isolation
  fixture. Remove one test that was a byte-identical copy of another.

`python -m kernelforge.cli` stays the single external CLI. Verified: every
production module imports, the 18 modules Hyperloom imports from kernelforge
resolve, all five subcommands respond, `scripts/tests/test_forge_e2e_gate.py`
passes, and the full suite is 4402 passed / 1 skipped.
