+++ src/kernelforge/fusion/validate.py
@@ -495,254 +495,6 @@ def _installed_package_dir(pkg: str) -> str:
     return str(Path(spec.origin).parent)
 
 
-# A fusion reaches the model by being called. Publishing it onto another module
-# is how that is arranged, so an assignment whose target is an attribute is the
-# thing to check; a plain local assignment is bookkeeping inside the new module.
-def _imported_module_aliases(tree: ast.Module) -> set[str]:
-    """Names in this file that refer to a module rather than a value."""
-    aliases: set[str] = set()
-    for node in ast.walk(tree):
-        if isinstance(node, ast.Import):
-            for alias in node.names:
-                aliases.add(alias.asname or alias.name.split(".")[0])
-        elif isinstance(node, ast.ImportFrom) and node.module:
-            for alias in node.names:
-                # `from pkg import mod as m` -- indistinguishable from importing a
-                # value here, and treating it as a module only widens the check.
-                aliases.add(alias.asname or alias.name)
-    return aliases
-
-
-def _published_attribute_names(source: str) -> set[str]:
-    """Attribute names this file installs onto another MODULE.
-
-    Only onto a module: `self.attn = ...` in an `__init__` is an instance
-    attribute and has nothing to do with publishing a kernel, and counting those
-    buries the real finding under every field the model assigns.
-    """
-    try:
-        tree = ast.parse(source)
-    except SyntaxError:
-        return set()
-    modules = _imported_module_aliases(tree)
-    names: set[str] = set()
-    for node in ast.walk(tree):
-        if not isinstance(node, ast.Assign):
-            continue
-        for target in node.targets:
-            if not isinstance(target, ast.Attribute):
-                continue
-            base = target.value
-            if isinstance(base, ast.Name) and base.id in modules:
-                names.add(target.attr)
-    return names
-
-
-def _top_level_names(source: str) -> set[str]:
-    """Names this module defines at its top level, excluding module metadata."""
-    try:
-        tree = ast.parse(source)
-    except SyntaxError:
-        return set()
-    names: set[str] = set()
-    for node in tree.body:
-        if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef, ast.ClassDef)):
-            names.add(node.name)
-        elif isinstance(node, ast.Assign):
-            # `short = long_name` is a synonym, not a second entry point. If what
-            # it points at is unreached that name is reported on its own, and
-            # counting the alias too would flag a wired fusion for keeping a
-            # spelling around.
-            if isinstance(node.value, ast.Name):
-                continue
-            for target in node.targets:
-                if isinstance(target, ast.Name):
-                    names.add(target.id)
-    return {n for n in names if not (n.startswith("__") and n.endswith("__"))}
-
-
-def _reads_by_owner(source: str) -> dict[str, set[str]]:
-    """Names read, keyed by the top-level definition that reads them.
-
-    Module-level reads are keyed by "" -- they run on import, so they count as
-    the framework reaching the name rather than as one new symbol citing another.
-    """
-    try:
-        tree = ast.parse(source)
-    except SyntaxError:
-        return {}
-    out: dict[str, set[str]] = {}
-    for node in tree.body:
-        owner = (
-            getattr(node, "name", "") if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef, ast.ClassDef)) else ""
-        )
-        bucket = out.setdefault(owner, set())
-        exported: set[str] = set()
-        for inner in ast.walk(node):
-            if isinstance(inner, ast.Assign):
-                exported |= {t.attr for t in inner.targets if isinstance(t, ast.Attribute)}
-            if isinstance(inner, ast.Attribute) and isinstance(inner.ctx, ast.Load):
-                bucket.add(inner.attr)
-            elif isinstance(inner, ast.Name) and isinstance(inner.ctx, ast.Load):
-                bucket.add(inner.id)
-            elif isinstance(inner, ast.Call):
-                # A name as a string is only a lookup inside a call --
-                # `getattr(mod, "op")`. In `__all__ = ["op"]` it is a listing,
-                # and counting it would let a module vouch for its own symbol.
-                bucket |= {
-                    arg.value for arg in inner.args if isinstance(arg, ast.Constant) and isinstance(arg.value, str)
-                }
-        # `other.op = mine.op` reads `op` on the right, and that read is the
-        # publish itself -- counting it would let a publisher vouch for its own
-        # symbol, which is the exact thing being tested for.
-        bucket -= exported
-    return out
-
-
-def unreached_fusion_symbols(
-    repo_root: str,
-    changed_files: list[str],
-    *,
-    pristine_dir: str = "",
-    _walk=None,
-) -> list[str]:
-    """Fusion symbols that nothing already in the model can reach.
-
-    Every gate the loop has can pass on a kernel that is never called. Compiling
-    proves it imports; parity and the microbench call it from the harness
-    directly; the serving smoke boots a server in which an unreferenced fusion is
-    simply inert. So a kernel can be authored, validated, kept and exported
-    without ever being on the model's execution path.
-
-    Two shapes of that have been seen and both are checked. One publishes the
-    kernel onto another module (`_attn.fused_op = ...`) that never looks the name
-    up. The other adds top-level definitions that nothing calls -- an audit of 27
-    landed fusions found one patch whose every hunk was a module-level insertion,
-    so no pre-existing function body was touched and there was no caller to be
-    had.
-
-    Reachability is transitive: a new definition cited only by another new
-    definition that is itself unreached does not count, or a self-contained
-    island of new code would look wired. Roots are the code that was already
-    there, which the framework calls by construction, plus module-level code,
-    which runs on import.
-
-    Without ``pristine_dir`` the "new definition" half degrades to the modules
-    whose names mark them author-created, since there is no baseline to diff
-    against. The published-attribute half does not need one.
-    """
-    root = Path(repo_root)
-    if not repo_root or not root.is_dir() or not changed_files:
-        return []
-    changed = [Path(f) if Path(f).is_absolute() else root / f for f in changed_files]
-    changed += _authored_modules_beside(changed, root, pristine_dir)
-    changed_set = {str(p) for p in changed}
-
-    published: set[str] = set()
-    introduced: set[str] = set()
-    for path in changed:
-        try:
-            text = path.read_text(encoding="utf-8", errors="ignore")
-        except OSError:
-            continue
-        published |= _published_attribute_names(text)
-        introduced |= _top_level_names(text) - _baseline_names(path, root, pristine_dir)
-
-    candidates = published | introduced
-    if not candidates:
-        return []
-
-    # Who reads what, per owning definition, across the tree.
-    walk = _walk or (lambda: root.rglob("*.py"))
-    reads: list[tuple[str, str, set[str]]] = []
-    for path in walk():
-        try:
-            text = path.read_text(encoding="utf-8", errors="ignore")
-        except OSError:
-            continue
-        if not any(name in text for name in candidates):
-            continue
-        for owner, names in _reads_by_owner(text).items():
-            hit = names & candidates
-            if hit:
-                reads.append((str(path), owner, hit))
-
-    # A definition is a root unless it is itself one of the new symbols.
-    reached: set[str] = set()
-    changing = True
-    while changing:
-        changing = False
-        for path_str, owner, names in reads:
-            owner_is_new = owner and path_str in changed_set and owner in candidates and owner not in reached
-            if owner_is_new:
-                continue
-            new_hits = names - reached
-            if new_hits:
-                reached |= new_hits
-                changing = True
-    # The question is whether the fusion is on the execution path, not whether
-    # every name it introduced is used. A fusion carries helpers the model is not
-    # supposed to call -- the eager reference the parity check compares against
-    # is the clearest case, and reporting it would fail a wired fusion for
-    # shipping the thing that proved it correct. One reached entry point means
-    # the model gets there.
-    if reached:
-        return []
-    return sorted(candidates)
-
-
-def _authored_modules_beside(changed: list[Path], root: Path, pristine_dir: str) -> list[Path]:
-    """Fused-kernel modules the author created next to a file it edited.
-
-    The caller knows the model source it asked for; it does not know what the
-    author put beside it. A fusion whose kernel lives in a new module and is
-    never called from the edited file is dead in exactly the way this checks
-    for, and passing only the edited file cannot see it.
-    """
-    known = {p.resolve() for p in changed}
-    found: list[Path] = []
-    for path in changed:
-        parent = path.parent
-        if not parent.is_dir():
-            continue
-        for sibling in sorted(parent.glob("*.py")):
-            if sibling.resolve() in known:
-                continue
-            name = sibling.name.lower()
-            if "fused" not in name and "fusion" not in name:
-                continue
-            if pristine_dir:
-                with contextlib.suppress(ValueError, OSError):
-                    snap = Path(pristine_dir) / sibling.resolve().relative_to(root.resolve())
-                    if snap.is_file():
-                        continue  # predates this run
-            found.append(sibling)
-            known.add(sibling.resolve())
-    return found
-
-
-def _baseline_names(path: Path, root: Path, pristine_dir: str) -> set[str]:
-    """Top-level names this file had before authoring.
-
-    A file with no snapshot is either author-created -- everything in it is new
-    -- or unknowable, in which case claiming everything is new would report the
-    whole module. Only the first is treated as new, by the same name test the
-    export path uses to decide what belongs to a fusion.
-    """
-    if pristine_dir:
-        with contextlib.suppress(ValueError, OSError):
-            snap = Path(pristine_dir) / path.resolve().relative_to(root.resolve())
-            if snap.is_file():
-                return _top_level_names(snap.read_text(encoding="utf-8", errors="ignore"))
-    name = path.name.lower()
-    if "fused" in name or "fusion" in name:
-        return set()
-    try:
-        return _top_level_names(path.read_text(encoding="utf-8", errors="ignore"))
-    except OSError:
-        return set()
-
-
 # The engine runs in a child process whose name does not contain the launcher's
 # command line, so a pkill written against the launcher leaves it holding the
 # card. Observed on this hardware: 283 of 288 GiB still allocated after the
