+++ src/kernelforge/tests/kernel_rewrite_controller/test_recovery.py
@@ -4,8 +4,6 @@
 from __future__ import annotations
 
 import json
-import os
-import subprocess
 from pathlib import Path
 
 from kernelforge.kernel_rewrite_controller import (
@@ -24,25 +22,7 @@ from kernelforge.knowledge.kernel_identity import (
     kernel_recipe_canonical_id,
 )
 from kernelforge.loop.reporting import BestResultPublisher
-
-_GIT_IDENTITY = {
-    "GIT_AUTHOR_NAME": "controller-test",
-    "GIT_AUTHOR_EMAIL": "controller-test@local",
-    "GIT_COMMITTER_NAME": "controller-test",
-    "GIT_COMMITTER_EMAIL": "controller-test@local",
-}
-
-
-def _git(repo: Path, *args: str) -> str:
-    result = subprocess.run(
-        ["git", *args],
-        cwd=repo,
-        env={**os.environ, **_GIT_IDENTITY},
-        capture_output=True,
-        text=True,
-        check=True,
-    )
-    return result.stdout.strip()
+from kernelforge.tests.kernel_rewrite_controller.conftest import _git
 
 
 def _prepared_workspace(tmp_path: Path):
