+++ src/kernelforge/tests/kernel_rewrite_controller/test_worktree.py
@@ -3,8 +3,6 @@
 
 from __future__ import annotations
 
-import os
-import subprocess
 from pathlib import Path
 
 import pytest
@@ -20,25 +18,7 @@ from kernelforge.knowledge.kernel_identity import (
     KernelRecipeIdentity,
     kernel_recipe_canonical_id,
 )
-
-_GIT_IDENTITY = {
-    "GIT_AUTHOR_NAME": "controller-test",
-    "GIT_AUTHOR_EMAIL": "controller-test@local",
-    "GIT_COMMITTER_NAME": "controller-test",
-    "GIT_COMMITTER_EMAIL": "controller-test@local",
-}
-
-
-def _git(repo: Path, *args: str) -> str:
-    completed = subprocess.run(
-        ["git", *args],
-        cwd=repo,
-        env={**os.environ, **_GIT_IDENTITY},
-        capture_output=True,
-        text=True,
-        check=True,
-    )
-    return completed.stdout.strip()
+from kernelforge.tests.kernel_rewrite_controller.conftest import _git
 
 
 def _source_repo(tmp_path: Path) -> tuple[Path, str]:
