+++ src/kernelforge/tests/test_tracker.py
@@ -11,7 +11,7 @@ from pathlib import Path
 
 import pytest
 
-from kernelforge.tracker import ExperimentTracker, Experiment
+from kernelforge.tracker import ExperimentTracker
 
 from kernelforge.conftest import SRC_ROOT
 
@@ -152,183 +152,6 @@ def test_log_iteration():
         loaded = tracker.get(exp.experiment_id)
         assert len(loaded.iterations) == 2
         assert loaded.iterations[1].wall_ms == 1.2
-
-
-def test_best_iteration():
-    with tempfile.TemporaryDirectory() as tmpdir:
-        tracker = ExperimentTracker(tmpdir)
-        exp = tracker.create(task_id="test", backend="ck", target_wall_ms=1.0)
-
-        # Iteration 1: passes SNR, mediocre perf
-        tracker.log_iteration(
-            exp.experiment_id,
-            snr_db=35.0,
-            wall_ms=2.0,
-            mean_case_speedup=1.0,
-        )
-        # Iteration 2: fails SNR — should be excluded from best
-        tracker.log_iteration(
-            exp.experiment_id,
-            snr_db=15.0,
-            wall_ms=0.5,
-            mean_case_speedup=4.0,
-        )
-        # Iteration 3: passes SNR, best perf
-        tracker.log_iteration(
-            exp.experiment_id,
-            snr_db=32.0,
-            wall_ms=1.1,
-            mean_case_speedup=2.0,
-        )
-
-        best = tracker.get_best(exp.experiment_id)
-        assert best is not None
-        assert best.iteration_id == 3
-        assert best.wall_ms == 1.1
-
-
-def test_reverted_iteration_cannot_become_best_or_meet_gate():
-    """Exclude a failed confirmation result from best-performance reporting."""
-    exp = Experiment(experiment_id="test", target_wall_ms=1.0)
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=1.1,
-        mean_case_speedup=1.0,
-        decision="KEEP",
-    )
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=0.9,
-        mean_case_speedup=2.0,
-        decision="REVERT",
-    )
-    assert exp.best_iteration().wall_ms == 1.1
-    assert exp.is_gate_met() is False
-
-
-def test_plateau_detection():
-    exp = Experiment(experiment_id="test")
-    # Add 3 authoritative scores with <2% variance.
-    exp.add_iteration(snr_db=30.0, wall_ms=1.00, mean_case_speedup=1.20)
-    exp.add_iteration(snr_db=31.0, wall_ms=0.99, mean_case_speedup=1.21)
-    exp.add_iteration(snr_db=30.5, wall_ms=0.995, mean_case_speedup=1.205)
-    assert exp.is_plateaued(n=3, threshold=0.02)
-
-    # Add iteration with significant improvement → no longer plateaued
-    exp.add_iteration(snr_db=30.0, wall_ms=0.80, mean_case_speedup=1.40)
-    assert not exp.is_plateaued(n=3, threshold=0.02)
-
-
-def test_gate_check():
-    exp = Experiment(experiment_id="test", target_wall_ms=1.0)
-    assert not exp.is_gate_met()
-
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=1.5,
-        mean_case_speedup=1.0,
-    )  # above target
-    assert not exp.is_gate_met()
-
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=0.9,
-        mean_case_speedup=2.0,
-    )  # below target
-    assert exp.is_gate_met()
-    assert exp.is_gate_met(exp.scoring_view())
-
-
-def test_mean_case_speedup():
-    exp = Experiment(experiment_id="test", baseline_wall_ms=2.0)
-    exp.add_iteration(snr_db=30.0, wall_ms=1.0, mean_case_speedup=2.0)
-    assert exp.best_mean_case_speedup() == 2.0
-
-
-def test_legacy_history_is_display_only():
-    exp = Experiment(
-        experiment_id="legacy",
-        baseline_wall_ms=2.0,
-        target_wall_ms=1.0,
-    )
-    legacy = exp.add_iteration(snr_db=35.0, wall_ms=1.0)
-
-    assert exp.best_iteration() is None
-    assert exp.legacy_best_iteration() is legacy
-    assert exp.display_best_iteration() is legacy
-    assert exp.display_speedup() == (2.0, "legacy raw ratio")
-    assert exp.is_gate_met() is False
-
-
-def test_display_prefers_authoritative_score_over_faster_legacy_raw_wall():
-    exp = Experiment(experiment_id="mixed", baseline_wall_ms=2.0)
-    exp.add_iteration(snr_db=35.0, wall_ms=0.5)
-    scored = exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=1.5,
-        mean_case_speedup=1.4,
-    )
-
-    assert exp.display_best_iteration() is scored
-    assert exp.display_speedup() == (1.4, "mean case speedup")
-
-
-def test_plateau_uses_authoritative_score_instead_of_raw_wall():
-    exp = Experiment(experiment_id="scored")
-    for wall_ms, speedup in (
-        (3.0, 1.200),
-        (1.0, 1.205),
-        (2.0, 1.210),
-    ):
-        exp.add_iteration(
-            snr_db=35.0,
-            wall_ms=wall_ms,
-            mean_case_speedup=speedup,
-        )
-
-    assert exp.is_plateaued()
-
-
-def test_plateau_does_not_fall_back_to_raw_wall_with_too_few_scores():
-    exp = Experiment(experiment_id="partially-scored")
-    exp.add_iteration(snr_db=35.0, wall_ms=1.000)
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=1.001,
-        mean_case_speedup=1.20,
-    )
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=1.002,
-        mean_case_speedup=1.21,
-    )
-
-    assert not exp.is_plateaued()
-
-
-def test_summary_table():
-    exp = Experiment(
-        experiment_id="test",
-        target_wall_ms=1.0,
-        baseline_wall_ms=2.0,
-    )
-    exp.add_iteration(
-        snr_db=35.0,
-        wall_ms=1.5,
-        mean_case_speedup=1.0,
-        decision="initial",
-    )
-    exp.add_iteration(
-        snr_db=33.0,
-        wall_ms=0.9,
-        mean_case_speedup=2.0,
-        decision="optimized",
-    )
-
-    table = exp.summary_table()
-    assert "Iter" in table
-    assert "0.900" in table
-    assert "Gate (1.0 ms): MET" in table
 
 
 def test_list_experiments():
