+++ src/kernelforge/tests/test_tracker_cov.py
@@ -77,14 +77,6 @@ def test_get_missing_raises(tmp_path):
         tracker.get("does_not_exist")
 
 
-def test_summary_delegates(tmp_path):
-    tracker = ExperimentTracker(tmp_path)
-    exp = tracker.create(task_id="t", backend="ck", target_wall_ms=1.0)
-    tracker.log_iteration(exp.experiment_id, snr_db=35.0, wall_ms=0.8)
-    summary = tracker.summary(exp.experiment_id)
-    assert "Iter" in summary
-
-
 def test_iteration_roundtrip_dict():
     it = Iteration(iteration_id=1, config={"BLOCK_M": 128}, snr_db=35.0, wall_ms=1.2)
     d = it.to_dict()
@@ -93,63 +85,6 @@ def test_iteration_roundtrip_dict():
     assert "notes" not in d
     restored = Iteration.from_dict({**d, "unknown": 1})
     assert restored.snr_db == 35.0
-
-
-def test_is_gate_met_no_target():
-    assert Experiment(experiment_id="e").is_gate_met() is False
-
-
-def test_effective_baseline_none():
-    exp = Experiment(experiment_id="e")
-    assert exp.effective_baseline_ms() is None
-    assert exp.best_mean_case_speedup() is None
-
-
-def test_consecutive_reverts():
-    exp = Experiment(experiment_id="e")
-    exp.add_iteration(snr_db=35.0, wall_ms=1.0, decision="KEEP")
-    exp.add_iteration(snr_db=35.0, wall_ms=1.1, decision="REVERT")
-    exp.add_iteration(snr_db=35.0, wall_ms=1.2, decision="REVERT")
-    assert exp.consecutive_reverts() == 2
-    # A KEEP breaks the streak.
-    exp.add_iteration(snr_db=35.0, wall_ms=0.9, decision="KEEP")
-    assert exp.consecutive_reverts() == 0
-
-
-def test_summary_table_legacy_data_is_unscored_and_not_plateaued():
-    exp = Experiment(experiment_id="e", target_wall_ms=0.5)
-    for wall in (1.00, 0.99, 0.995):
-        exp.add_iteration(snr_db=35.0, wall_ms=wall)
-    table = exp.summary_table()
-    assert "Gate (" not in table
-    assert "PLATEAUED" not in table
-
-
-def test_uses_authoritative_scoring_requires_a_per_case_score():
-    """Raw wall time alone is display-only history, never an authoritative score."""
-    scored = Experiment(experiment_id="scored")
-    scored.add_iteration(snr_db=35.0, wall_ms=1.0, mean_case_speedup=1.2)
-    legacy = Experiment(experiment_id="legacy")
-    legacy.add_iteration(snr_db=35.0, wall_ms=1.0)
-
-    assert scored.uses_authoritative_scoring() is True
-    assert legacy.uses_authoritative_scoring() is False
-    # The legacy raw best is offered only while scoring is unauthoritative.
-    assert scored.legacy_best_iteration() is None
-    assert legacy.legacy_best_iteration() is not None
-
-
-def test_summary_table_announces_plateau_and_reverted_changes():
-    exp = Experiment(experiment_id="e", baseline_wall_ms=2.0)
-    for speedup, wall in ((1.100, 1.00), (1.110, 0.99), (1.105, 0.995)):
-        exp.add_iteration(snr_db=35.0, wall_ms=wall, mean_case_speedup=speedup)
-    exp.changes_reverted = ["unrolled epilogue", "widened lds tile"]
-
-    table = exp.summary_table()
-
-    assert exp.is_plateaued() is True
-    assert "PLATEAUED" in table
-    assert "Reverted: unrolled epilogue, widened lds tile" in table
 
 
 def test_experiment_to_from_dict_roundtrip():
